feat: Add react-query and axios configurations
This commit adds the necessary configurations for using react-query and axios in the project. It includes the creation of a `react-query.ts` file to initialize the `QueryClient` and a `axios.ts` file to create an instance of axios with the base URL from the environment variables. These configurations will enable the usage of react-query for data fetching and axios for making HTTP requests in the application.